Code commited by karthick

+++ src/main/java/com/rubix/hrm/repository/AttendanceRepository.java
@@ -7,6 +7,6 @@ import org.springframework.stereotype.Repository;
 import com.rubix.hrm.models.Attendance;
 
 @Repository
-public interface AttendanceRepository extends JpaRepository<Attendance, Integer> {
+public interface AttendanceRepository extends JpaRepository<Attendance, Long> {
 
 }
